moved some tests around
